#50 - rename test package name from org.apache to com.github
