+++ src/test/java/com/github/cameltooling/lsp/internal/completion/CamelComponentOptionBooleanValuesTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.tools.lsp.internal.completion;
+package com.github.cameltooling.lsp.internal.completion;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -23,20 +23,20 @@ import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
-import org.apache.camel.tools.lsp.internal.AbstractCamelLanguageServerTest;
 import org.eclipse.lsp4j.CompletionItem;
 import org.eclipse.lsp4j.CompletionList;
 import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.jsonrpc.messages.Either;
 import org.junit.Test;
 
+import com.github.cameltooling.lsp.internal.AbstractCamelLanguageServerTest;
 import com.github.cameltooling.lsp.internal.CamelLanguageServer;
 
-public class CamelComponentOptionEnumerationValuesTest extends AbstractCamelLanguageServerTest {
+public class CamelComponentOptionBooleanValuesTest extends AbstractCamelLanguageServerTest {
 
 	@Test
-	public void testProvideEnumValues() throws Exception {
-		testProvideCamelOptions("<from uri=\"timer:timerName?exchangePattern=\" xmlns=\"http://camel.apache.org/schema/blueprint\"></from>\n", 0, 43);
+	public void testProvideBooleanValues() throws Exception {
+		testProvideCamelOptions("<from uri=\"timer:timerName?fixedRate=\" xmlns=\"http://camel.apache.org/schema/blueprint\"></from>\n", 0, 37);
 	}
 
 	private void testProvideCamelOptions(String textTotest, int line, int character) throws URISyntaxException, InterruptedException, ExecutionException {
@@ -45,14 +45,8 @@ public class CamelComponentOptionEnumerationValuesTest extends AbstractCamelLang
 		CompletableFuture<Either<List<CompletionItem>, CompletionList>> completions = getCompletionFor(camelLanguageServer, new Position(line, character));
 
 		assertThat(completions.get().getLeft()).contains(
-				new CompletionItem("InOnly"),
-				new CompletionItem("RobustInOnly"),
-				new CompletionItem("InOut"),
-				new CompletionItem("InOptionalOut"),
-				new CompletionItem("OutOnly"),
-				new CompletionItem("RobustOutOnly"),
-				new CompletionItem("OutIn"),
-				new CompletionItem("OutOptionalIn"));
+				new CompletionItem("true"),
+				new CompletionItem("false"));
 	}
 
 }
